+++ projects/start-os/web/ui/src/app/routes/portal/modals/qr.component.ts
@@ -1,12 +1,12 @@
 import { Component } from '@angular/core'
 import { TuiDialogContext } from '@taiga-ui/core'
 import { injectContext } from '@taiga-ui/polymorpheus'
-import { QrCodeModule } from 'ng-qrcode'
+import { QRComponent } from 'src/app/components/qr.component'
 
 @Component({
   selector: 'qr',
-  template: '<qr-code [value]="context.data" size="350"></qr-code>',
-  imports: [QrCodeModule],
+  template: '<app-qr [value]="context.data" />',
+  imports: [QRComponent],
 })
 export class QRModal {
   readonly context = injectContext<TuiDialogContext<void, string>>()